Basic Giveaway Application
A project i made for our school club, where you can make a raffle with as many winners and reserve winners as you want, by reading the names one by one in a text file on your computer. It was a pastime project but i had a lot of fun while doing it and i took as much care as possible.